backend/requirements.txt
fastapi
motor
pydantic
passlib[bcrypt]
python-jose
uvicorn
python-dotenv
